port over remaining types and delete js files
